moved obsolete tests and classes to sandbox package under test folder
