+++ src/test/java/com/mindlinksoft/recruitment/mychat/AllTests.java
@@ -4,6 +4,9 @@ import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
 
+import sandbox.ConversationExporterTests;
+import sandbox.FiltererTest;
+
 @RunWith(Suite.class)
 @SuiteClasses({ CommandLineArgumentParserTests.class, ConversationExporterTests.class, FiltererTest.class,
 		MainCLITests.class })
